requirements.txt
Jinja2==2.9.1
PyYAML==3.12
jmespath==0.9.0
jupyter
mistune
termcolor
watchdog==0.8.3
nbconvert==5.0.0
python-slugify==1.2.1
